feat(tests): add core tests project
