+++ src/backend/Auth/Actions/UserSettingsActions/PatchUserDetails/PatchUserDetailsPayload.cs
@@ -1,0 +1,8 @@
+﻿namespace Auth.Actions.UserSettingsActions.PatchUserDetails
+{
+    public class PatchUserDetailsPayload
+    {
+        public string FirstName { get; set; } = default!;
+        public string LastName { get; set; } = default!;
+    }
+}
