--- src/backend/AuthService/Actions/UserSettingsActions/PatchUserDetails/PatchUserDetailsPayload.cs
@@ -1,8 +1,0 @@
-﻿namespace AuthService.Actions.UserSettingsActions.PatchUserDetails
-{
-    public class PatchUserDetailsPayload
-    {
-        public string FirstName { get; set; } = default!;
-        public string LastName { get; set; } = default!;
-    }
-}
