+++ src/backend/AuthTests/AuthTests.csproj
@@ -25,6 +25,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<ProjectReference Include="..\Auth\Auth.csproj" />
 		<ProjectReference Include="..\TestsBase\TestsBase.csproj" />
 	</ItemGroup>
