+++ src/backend/Donate/DonateModule.cs
@@ -1,0 +1,12 @@
+﻿using Autofac;
+
+namespace Donate
+{
+    public class DonateServiceModule : Module
+    {
+        protected override void Load(ContainerBuilder builder)
+        {
+
+        }
+    }
+}
