+++ src/backend/Donate/RemoveThisClass.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Donate
+{
+    public class RemoveThisClass
+    {
+        public int TestMe(int a, int b) { return a + b; }
+    }
+}
