--- src/backend/DonateService/DonateServiceModule.cs
@@ -1,12 +1,0 @@
-﻿using Autofac;
-
-namespace DonateService
-{
-    public class DonateServiceModule : Module
-    {
-        protected override void Load(ContainerBuilder builder)
-        {
-
-        }
-    }
-}
