--- src/backend/DonateService/RemoveThisClass.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace DonateService
-{
-    public class RemoveThisClass
-    {
-        public int TestMe(int a, int b) { return a + b; }
-    }
-}
