--- src/backend/DonateServiceTests/RemoveThisClassTests.cs
@@ -1,15 +1,0 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-
-namespace DonateService.Tests
-{
-    [TestClass()]
-    public class RemoveThisClassTests
-    {
-        [TestMethod()]
-        public void TestMeTest()
-        {
-            var abc = new RemoveThisClass();
-            Assert.AreEqual(3, abc.TestMe(1, 2));
-        }
-    }
-}
