+++ src/backend/DonateTests/DonateTests.csproj
@@ -25,6 +25,7 @@
 	</ItemGroup>
 
 	<ItemGroup>
+		<ProjectReference Include="..\Donate\Donate.csproj" />
 		<ProjectReference Include="..\TestsBase\TestsBase.csproj" />
 	</ItemGroup>
