+++ src/backend/InGoodHands.sln
@@ -5,7 +5,7 @@ VisualStudioVersion = 17.0.31903.59
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "WebApi", "WebApi\WebApi.csproj", "{585775DA-CB84-4E68-9910-D02FBBB8B2B7}"
 EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Services", "Services", "{B39D7A08-C7A7-437E-BD2E-652E5A503B45}"
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Features", "Features", "{B39D7A08-C7A7-437E-BD2E-652E5A503B45}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Presentation", "Presentation", "{FAFB8B32-7214-48AF-9331-59EA201AA75A}"
 EndProject
@@ -13,17 +13,17 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Core", "Core\Core.csproj", 
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Worker", "Worker\Worker.csproj", "{2322E411-F587-4C24-9FC1-83ABF168853C}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthService", "AuthService\AuthService.csproj", "{CC5910D1-5FBF-4B8C-83FA-A40D5A5ACD45}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Auth", "Auth\Auth.csproj", "{CC5910D1-5FBF-4B8C-83FA-A40D5A5ACD45}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Tests", "Tests", "{588F1E45-AB8A-4663-95C0-B99154122AE4}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TestsBase", "TestsBase\TestsBase.csproj", "{17E086CD-5571-49B1-BAE4-B52958394489}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthServiceTests", "AuthServiceTests\AuthServiceTests.csproj", "{26D081DE-7E04-43A9-85DF-E348253BD505}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "AuthTests", "AuthTests\AuthTests.csproj", "{26D081DE-7E04-43A9-85DF-E348253BD505}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DonateService", "DonateService\DonateService.csproj", "{52603855-00A3-473E-B890-C99D9C344D09}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Donate", "Donate\Donate.csproj", "{52603855-00A3-473E-B890-C99D9C344D09}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DonateServiceTests", "DonateServiceTests\DonateServiceTests.csproj", "{95563EC6-EED2-4A9B-8009-54AE0ABCB8D1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DonateTests", "DonateTests\DonateTests.csproj", "{95563EC6-EED2-4A9B-8009-54AE0ABCB8D1}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "CoreTests", "CoreTests\CoreTests.csproj", "{D3E34DEC-526E-485F-862C-E08939AFCD0F}"
 EndProject
